refactor: packages names, urls, TaskModel take off constructor
