+++ src/main/java/cvds/todo/backend/controller/TaskController.java
@@ -1,7 +1,8 @@
 package cvds.todo.backend.controller;
 
 import cvds.todo.backend.model.TaskModel;
-import cvds.todo.backend.service.TaskService;
+import cvds.todo.backend.services.TaskService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,26 +14,18 @@ import java.util.UUID;
  * Proporciona endpoints para crear, leer, actualizar y eliminar tareas.
  */
 @RestController
-@RequestMapping("/task")
-public class Controller {
+@RequestMapping("/tasks")
+public class TaskController {
 
-    private final TaskService taskService;
-
-    /**
-     * Constructor que inyecta el servicio de tareas.
-     *
-     * @param taskService Servicio para gestionar las tareas.
-     */
-    public Controller(TaskService taskService) {
-        this.taskService = taskService;
-    }
+    @Autowired
+    private TaskService taskService;
 
     /**
      * Obtener todas las tareas.
      *
      * @return Lista de todas las tareas.
      */
-    @GetMapping("/all")
+    @GetMapping("/")
     public ResponseEntity<List<TaskModel>> getAllTasks() {
         List<TaskModel> tasks = taskService.getAllTasks();
         return ResponseEntity.ok(tasks);
@@ -56,10 +49,10 @@ public class Controller {
      * @param task Modelo de tarea enviado en el cuerpo de la solicitud.
      * @return El UUID de la nueva tarea creada.
      */
-    @PostMapping("/new")
-    public ResponseEntity<UUID> createTask(@RequestBody TaskModel task) {
-        TaskModel newTask = new TaskModel(UUID.randomUUID(), task);
-        return ResponseEntity.ok(newTask.getId());
+    @PostMapping("/")
+    public ResponseEntity<String> createTask(@RequestBody TaskModel task) {
+        taskService.createTask(task);
+        return ResponseEntity.ok("Ok");
     }
 
     /**
@@ -69,7 +62,7 @@ public class Controller {
      * @param task Modelo de tarea actualizado.
      * @return La tarea actualizada.
      */
-    @PutMapping("/{id}")
+    @PatchMapping("/{id}")
     public ResponseEntity<TaskModel> updateTask(@PathVariable("id") int id, @RequestBody TaskModel task) {
         TaskModel updatedTask = taskService.updateTask(id, task);
         return ResponseEntity.ok(updatedTask);
