+++ src/main/java/cvds/todo/backend/interfeces/TasksService.java
@@ -1,4 +1,4 @@
-package cvds.todo.backend.interfece;
+package cvds.todo.backend.interfeces;
 
 import cvds.todo.backend.model.TaskModel;
